requirements.txt
numpy>=1.26.4
pandas>=2.2.2
matplotlib>=3.10.0
scikit-learn>=1.5.2
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
ngboost>=0.5.8
pgbm>=2.2.0
gpboost>=1.6.1
mapie>=0.6.0
puncc>=0.8.0
pytorch-tabnet>=4.1.0
optuna>=3.6.0
optunahub>=0.4.0
openpyxl>=3.1.0
scipy>=1.10.0
torch>=2.0.0
pillow>=10.0.0
xlsxwriter>=3.0.0
